refactor: move JSON utils to use them in app
