+++ src/app/app.config.server.ts
@@ -1,8 +1,8 @@
 import { ApplicationConfig, mergeApplicationConfig } from '@angular/core'
 import { provideServerRendering } from '@angular/platform-server'
 import { appConfig } from './app.config'
-import { JsonFetcher } from './common/json-fetcher/json-fetcher'
-import { LocalJsonFetcherService } from './common/json-fetcher/local-json-fetcher.service'
+import { JsonFetcher } from '@/app/common/json/json-fetcher'
+import { LocalJsonFetcherService } from '@/app/common/json/local-json-fetcher.service'
 import { provideServerRouting } from '@angular/ssr'
 import { serverRoutes } from '@/app/app.routes.server'
