+++ src/app/app.routes.server.ts
@@ -5,13 +5,16 @@ import {
   PROJECTS_DIR,
   PUBLIC_DIR,
 } from '@/app/common/directories'
-import { basename, join } from 'path'
+import { join } from 'path'
 import { cwd } from 'process'
-import { readdir } from 'fs/promises'
+import { PROJECTS_PATH } from '@/app/common/routing/paths'
+import { SLUG_PARAM } from '@/app/common/routes-params'
+import { listJsonFilesInDirectory } from '@/app/common/json/json-file-utils'
+import { removeJsonExtension } from '@/app/common/json/json-extension-utils'
 
 export const serverRoutes: ServerRoute[] = [
   {
-    path: 'projects/:slug',
+    path: `${PROJECTS_PATH}/:${SLUG_PARAM}`,
     renderMode: RenderMode.Prerender,
     getPrerenderParams: () => slugsInDirectory(PROJECTS_DIR),
   },
@@ -24,6 +27,8 @@ export const serverRoutes: ServerRoute[] = [
 const slugsInDirectory = async (
   directory: string,
 ): Promise<Record<string, string>[]> =>
-  (await readdir(join(cwd(), PUBLIC_DIR, CONTENTS_DIR, directory))).map(
-    (file) => ({ slug: basename(file, '.json') }),
-  )
+  (
+    await listJsonFilesInDirectory(
+      join(cwd(), PUBLIC_DIR, CONTENTS_DIR, directory),
+    )
+  ).map((file) => ({ slug: removeJsonExtension(file) }))
