+++ src/app/projects/project-detail-page/project-detail-page-resolver.service.ts
@@ -2,9 +2,9 @@ import { ActivatedRouteSnapshot, ResolveFn } from '@angular/router'
 import { EMPTY, tap } from 'rxjs'
 import { ProjectDetail } from '../project'
 import { ProjectsService } from '../projects.service'
-import { SLUG_PARAM } from '../projects.routes-params'
 import { NavigatorService } from '../../common/routing/navigator.service'
 import { Injectable } from '@angular/core'
+import { SLUG_PARAM } from '@/app/common/routes-params'
 
 @Injectable()
 export class ProjectDetailPageResolver {
