+++ src/app/projects/projects.routes.ts
@@ -2,7 +2,7 @@ import { Route, Routes } from '@angular/router'
 import { ProjectDetailPageComponent } from './project-detail-page/project-detail-page.component'
 import { ProjectDetailPageResolver } from './project-detail-page/project-detail-page-resolver.service'
 import { inject } from '@angular/core'
-import { SLUG_PARAM } from './projects.routes-params'
+import { SLUG_PARAM } from '@/app/common/routes-params'
 import { ProjectDetailRouteData } from './project-detail-page/projects-routes-data'
 import { RouteDataResolver } from '../common/routing/route-data-resolver'
 import { ProjectsService } from './projects.service'
