+++ src/app/projects/projects.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core'
-import { JsonFetcher } from '../common/json-fetcher/json-fetcher'
+import { JsonFetcher } from '@/app/common/json/json-fetcher'
 import { PROJECTS_DIR } from '../common/directories'
 import { from, Observable } from 'rxjs'
 import { ProjectDetail, ProjectListItem } from './project'
@@ -9,14 +9,10 @@ export class ProjectsService {
   constructor(private readonly _jsonFetcher: JsonFetcher) {}
 
   async getListItems(): Promise<readonly ProjectListItem[]> {
-    return this._jsonFetcher.fetch<readonly ProjectListItem[]>(
-      `${PROJECTS_DIR}.json`,
-    )
+    return this._jsonFetcher.fetch(PROJECTS_DIR)
   }
 
   getDetail(slug: string): Observable<ProjectDetail> {
-    return from(
-      this._jsonFetcher.fetch<ProjectDetail>(PROJECTS_DIR, `${slug}.json`),
-    )
+    return from(this._jsonFetcher.fetch<ProjectDetail>(PROJECTS_DIR, slug))
   }
 }
